+++ inc/define.php
@@ -406,6 +406,7 @@ $CFG_GLPI["rulecollections_types"]        = ['RuleImportAssetCollection',
     'RuleRightCollection',
     'RuleSoftwareCategoryCollection',
     'RuleTicketCollection',
+    'RuleChangeCollection',
                                                   //'RuleImportComputerCollection', //deprecated
     'RuleAssetCollection'
 ];
