+++ install/empty_data.php
@@ -7682,6 +7682,18 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
                 'profiles_id' => self::PROFILE_READ_ONLY,
                 'name' => 'database',
                 'rights' => READ,
+            ], [
+                'profiles_id' => self::PROFILE_SUPER_ADMIN,
+                'name' => 'rule_change',
+                'rights' => READ | UPDATE | CREATE | PURGE | RuleTicket::PARENT,
+            ], [
+                'profiles_id' => self::PROFILE_ADMIN,
+                'name' => 'rule_change',
+                'rights' => READ,
+            ], [
+                'profiles_id' => self::PROFILE_READ_ONLY,
+                'name' => 'rule_change',
+                'rights' => READ,
             ],
         ];
