+++ src/RuleChangeCollection.php
@@ -31,27 +31,14 @@
  * ---------------------------------------------------------------------
  */
 
-class RuleTicketCollection extends RuleCommonITILObjectCollection
+class RuleChangeCollection extends RuleCommonITILObjectCollection
 {
-   // From RuleCollection
-    public static $rightname    = 'rule_ticket';
-    public $menu_option         = 'ticket';
+    // From RuleCollection
+    public static $rightname                        = 'rule_change';
+    public $menu_option                             = 'change';
 
     public function getTitle()
     {
-        return __('Business rules for tickets');
-    }
-
-    /**
-     * @see RuleCollection::prepareInputDataForProcess()
-     **/
-    public function prepareInputDataForProcess($input, $params)
-    {
-
-       // Pass x-priority header if exists
-        if (isset($input['_head']['x-priority'])) {
-            $input['_x-priority'] = $input['_head']['x-priority'];
-        }
-        return parent::prepareInputDataForProcess($input, $params);
+        return __('Business rules for changes');
     }
 }
